% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simulate.R
\name{generate_outbreaks}
\alias{generate_outbreaks}
\title{Function to generate n_samples * n_part stochastic outbreaks for a given model output}
\usage{
generate_outbreaks(
  model_run,
  model,
  list_specs,
  list_data,
  n_part,
  aggreg_year = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{model_run}{mcstate_pmcmc object, output from seirvodin::run_model().}

\item{model}{dust model.}

\item{list_specs}{list containing the specification of the simulations, output from seirvodin::specs_simulations().}

\item{list_data}{list containing all data needed for the simulations, expecting 7 elements: `ref_m`: contact matrix between age groups (matrix); `ref_d`: degree of connectivity between regions (matrix); `new_birth`: number of daily births per region (matrix); `mean_import_per_reg`: number of annual import by year and region (matrix);  `year_per_age`: duration of each age group (vector); `dt_vacc`: vaccine coverage by year and age (data table with at least 5 columns: years, region, coverage, yob (for year of birth), and dose); `N`: number of inhabitants at t0 by region and age (matrix).}

\item{n_part}{Number of simulation per sample.}

\item{aggreg_year}{boolean: whether to aggregate the number of cases simulated by year.}

\item{verbose}{boolean: print the simulation step for each sample.}
}
\value{
3-dimensional array containing the number of cases by compartment (stratified by vaccine status, age and region), simulation, and day (or year if aggreg_year is TRUE)
}
\description{
Function to generate n_samples * n_part stochastic outbreaks for a given model output
}
