% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_fit.R
\name{run_model}
\alias{run_model}
\title{Run odin.dust model}
\usage{
run_model(
  list_specs,
  list_data,
  init,
  list_prior,
  list_min_max,
  model,
  proposal_matrix = NULL
)
}
\arguments{
\item{list_specs}{list containing the specification of the model fitting process, output from seirvodin::specs_run().}

\item{list_data}{list containing all data needed for model fitting, expecting 8 elements: `ref_m`: contact matrix between age groups (matrix); `ref_d`: degree of connectivity between regions (matrix); `new_birth`: number of daily births per region (matrix); `mean_import_per_reg`: number of annual import by year and region (matrix);  `year_per_age`: duration of each age group (vector); `dt_vacc`: vaccine coverage by year and age (data table with at least 5 columns: years, region, coverage, yob (for year of birth), and dose); `N`: number of inhabitants at t0 by region and age (matrix);`dt_case`: data table containing the number of cases by date, age and region (three columns: date (numeric starting at 0); cases; and population (character containing unique id for a strata of vaccine status, age, and region))}

\item{init}{Vector of initial value for each parameter}

\item{list_min_max}{list containing two vectors: `min` and `max`, each list contains the minimum (and maximum) value for each parameter}

\item{model}{dust model.}

\item{proposal_matrix}{Proposal matrix to run the MCMC, leave to NULL to start with default matrix}

\item{prior}{List of prior function for each parameter (for parameters without prior, the element should be  NULL)}
}
\value{
An `mcstate_pmcmc` object, containing all information on the model fit.
}
\description{
Run odin.dust model
}
