% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirvodin.R
\docType{package}
\name{seirvodin-package}
\alias{seirvodin}
\alias{seirvodin-package}
\title{seirvodin: SEIRV model stratified by age and region implemented in odin_dust}
\description{
Contains an SEIRV model stratified by age and region, implemented in odin_dust, along with functions to fit the model to case data, and generate stochastic simulations.
}
\author{
\strong{Maintainer}:  alxsrobert \email{alexis.robert@lshtm.ac.uk}

}
\keyword{internal}
