% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_simulate.R
\name{specs_simulations}
\alias{specs_simulations}
\title{Set specifications of the simulation run}
\usage{
specs_simulations(
  N_year,
  year_start,
  waning,
  burnin,
  n_samples,
  nowane = FALSE,
  deterministic = FALSE
)
}
\arguments{
\item{N_year}{Number of years the simulations run for.}

\item{year_start}{starting year of the simulations.}

\item{waning}{does the model include waning of immunity? Should be one of "no", "since_eli", "since_vax", or "early".}

\item{burnin}{Numeric: duration of the burnin period}

\item{n_samples}{Numeric: Number of samples used to generate the simulations}

\item{nowane}{boolean: if set to TRUE, v_leak will be set to 0, used to quantify the impact of waning on the overall number of cases}

\item{deterministic}{boolean: set to yes to run a deterministic version of the model}
}
\value{
List of specifications
}
\description{
Set specifications of the simulation run
}
